requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
attrs==24.2.0
discord.py==2.5.2
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
pickleDB==0.9.2
propcache==0.2.0
py-memoize==3.1.1
python-dotenv==1.0.1
yarl==1.15.5
